updated def factorial

+++ src/main/scala/problems/math.scala
@@ -15,9 +15,9 @@ def isPrime(n: Int): Boolean =
 
   loop(2)
 
-def factorial(n: Int): Int =
+def factorial(n: Int): Long =
   @tailrec
-  def loop(acc: Int, k: Int): Int =
+  def loop(acc: Long, k: Int): Long =
     if k <= 0 then
       acc
     else
